backend/requirements.txt
asgiref==3.5.2
Django==4.0.4
djoser==2.1.0
django-cors-headers==3.12.0
djangorestframework==3.13.1
Pillow==9.1.0
pytils==0.4.1
pytz==2022.1
slugify==0.0.1
sqlparse==0.4.2
requests==2.27.1